Move audit and file check to their own package
